requirements.txt
flask==3.0.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.26.2
google-generativeai==0.3.1
requests==2.31.0
